2016-3-23 微信支付回调 by张佳

+++ GKPrison/app/src/main/java/com/gkzxhn/gkprison/userport/bean/Order.java
@@ -10,7 +10,7 @@ public class Order {
     private int jail_id;
     private List<line_items_attributes> line_items_attributes;
     private String created_at;
-    private int amount;
+    private float amount;
     private int family_id;
 
     public int getFamily_id() {
@@ -26,7 +26,7 @@ public class Order {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(float amount) {
         this.amount = amount;
     }
